+++ packages/clients/textLocator/vite.config.js
@@ -2,4 +2,10 @@ import { getClientConfig } from '../../../viteConfigs'
 
 export default getClientConfig({
   base: '',
+  build: {
+    lib: {
+      entry: 'polar-client.ts',
+      name: 'POLAR TextLocator Client',
+    },
+  },
 })